+++ src/main/java/com/ejemplo/spring/facturacion/bean/ComprobanteBean.java
@@ -6,14 +6,14 @@ public class ComprobanteBean implements Serializable
 {
 	private static final long serialVersionUID = 1L;
 	
-	private Long ID;
-	private Double numeroComprobante;
+	private Integer ID;
+	private Integer numeroComprobante;
 	private String nombreCliente;
 	private String apellidoPaterno;
 	private String apellidoMaterno;
 	private String DNI;
 	private String RUC;
-	private Double montototal;
+	private Float montototal;
 	private String estado;
 	private String sede;
 	
@@ -30,10 +30,10 @@ public class ComprobanteBean implements Serializable
 	public void setEstado(String estado) {
 		this.estado = estado;
 	}
-	public Long getID() {
+	public Integer getID() {
 		return ID;
 	}
-	public void setID(Long iD) {
+	public void setID(Integer iD) {
 		ID = iD;
 	}
 	public String getDNI() {
@@ -67,16 +67,16 @@ public class ComprobanteBean implements Serializable
 		this.apellidoMaterno = apellidoMaterno;
 	}
 
-	public Double getNumeroComprobante() {
+	public Integer getNumeroComprobante() {
 		return numeroComprobante;
 	}
-	public void setNumeroComprobante(Double numeroComprobante) {
+	public void setNumeroComprobante(Integer numeroComprobante) {
 		this.numeroComprobante = numeroComprobante;
 	}
-	public Double getMontototal() {
+	public Float getMontototal() {
 		return montototal;
 	}
-	public void setMontototal(Double montototal) {
+	public void setMontototal(Float montototal) {
 		this.montototal = montototal;
 	}
 	@Override
