+++ src/main/java/com/ejemplo/spring/facturacion/bean/LibroBean.java
@@ -6,15 +6,15 @@ public class LibroBean implements Serializable
 {
 	private static final long serialVersionUID = 1L;
 	
-	private Long ID;
+	private Integer ID;
 	private String nombreLibro;
-	private Double precioLibro;
+	private Float precioLibro;
 	
 	
-	public Long getID() {
+	public Integer getID() {
 		return ID;
 	}
-	public void setID(Long iD) {
+	public void setID(Integer iD) {
 		ID = iD;
 	}
 	public String getNombreLibro() {
@@ -23,10 +23,10 @@ public class LibroBean implements Serializable
 	public void setNombreLibro(String nombreLibro) {
 		this.nombreLibro = nombreLibro;
 	}
-	public Double getPrecioLibro() {
+	public Float getPrecioLibro() {
 		return precioLibro;
 	}
-	public void setPrecioLibro(Double precioLibro) {
+	public void setPrecioLibro(Float precioLibro) {
 		this.precioLibro = precioLibro;
 	}
